+++ admin/handlers/types-templates.go
@@ -1,7 +1,7 @@
 package handlers
 
 import (
-	"github.com/jmpsec/osctrl/carves"
+	"github.com/jmpsec/osctrl/pkg/carves"
 	"github.com/jmpsec/osctrl/pkg/environments"
 	"github.com/jmpsec/osctrl/pkg/nodes"
 	"github.com/jmpsec/osctrl/pkg/queries"
