+++ tls/handlers/handlers.go
@@ -5,8 +5,8 @@ import (
 	"strconv"
 	"time"
 
-	"github.com/jmpsec/osctrl/carves"
 	"github.com/jmpsec/osctrl/metrics"
+	"github.com/jmpsec/osctrl/pkg/carves"
 	"github.com/jmpsec/osctrl/pkg/environments"
 	"github.com/jmpsec/osctrl/pkg/logging"
 	"github.com/jmpsec/osctrl/pkg/nodes"
